+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,15 +2,15 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_201315: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_201114: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_201317: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_201327: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_201313: (a: number, b: number, c: number, d: number) => void;
-export const __wasm_bindgen_func_elem_194491: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_194502: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_196616: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_196636: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_201263: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_201064: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_201265: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_201277: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_194441: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_194452: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_201267: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_196566: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_196586: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;